Commit small test json files for each test instead of the large ones (#509)
* stop compressing json in ssv tests

* commit small test json files instead of large ones

* merge

* bug fix

* gosec

* remove tests.json in qbft

* clear tests folder before generating tests

* update logging and test files permission

* lint

* lint

* add nosec

* revert small test files permisison to 400

* pretty json

+++ types/spectest/generate/main.go
@@ -18,6 +18,7 @@ import (
 
 func main() {
 	clearStateComparisonFolder()
+	clearTestsFolder()
 
 	all := map[string]spectest.SpecTest{}
 	for _, t := range spectest.AllTests {
@@ -27,19 +28,40 @@ func main() {
 		}
 		all[n] = t
 	}
+	log.Printf("found %d tests\n", len(all))
+	if len(all) != len(spectest.AllTests) {
+		log.Fatalf("did not generate all tests\n")
+	}
 
-	byts, err := json.Marshal(all)
+	// write small json files for each test
+	// try to create directory if it doesn't exist
+	_, basedir, _, ok := runtime.Caller(0)
+	if !ok {
+		log.Fatalf("no caller info")
+	}
+	testsDir := filepath.Join(strings.TrimSuffix(basedir, "main.go"), "tests")
+	if err := os.MkdirAll(testsDir, 0700); err != nil && !os.IsExist(err) {
+		panic(err.Error())
+	}
+	for name, test := range all {
+		byts, err := json.MarshalIndent(test, "", "  ")
+		if err != nil {
+			panic(err.Error())
+		}
+		name = strings.ReplaceAll(name, " ", "_")
+		name = strings.ReplaceAll(name, "*", "")
+		name = "tests/" + name
+		writeJson(name, byts)
+	}
+
+	// write large tests.json file
+	byts, err := json.MarshalIndent(all, "", "  ")
 	if err != nil {
 		panic(err.Error())
 	}
+	writeJson("tests", byts)
 
-	if len(all) != len(spectest.AllTests) {
-		panic("did not generate all tests\n")
-	}
-
-	fmt.Printf("found %d tests\n", len(all))
-	writeJson(byts)
-
+	// write state comparison json files
 	for _, testF := range spectest.AllTests {
 		// generate post state comparison
 		writeJsonStateComparison(testF.TestName(), reflect.TypeOf(testF).String(), testF)
@@ -52,6 +74,22 @@ func clearStateComparisonFolder() {
 		panic("no caller info")
 	}
 	dir := filepath.Join(strings.TrimSuffix(basedir, "main.go"), "state_comparison")
+
+	if err := os.RemoveAll(dir); err != nil {
+		panic(err.Error())
+	}
+
+	if err := os.Mkdir(dir, 0700); err != nil {
+		panic(err.Error())
+	}
+}
+
+func clearTestsFolder() {
+	_, basedir, _, ok := runtime.Caller(0)
+	if !ok {
+		panic("no caller info")
+	}
+	dir := filepath.Join(strings.TrimSuffix(basedir, "main.go"), "tests")
 
 	if err := os.RemoveAll(dir); err != nil {
 		panic(err.Error())
@@ -96,7 +134,8 @@ func scDir(testType string) string {
 	scDir := comparable2.GetSCDir(basedir, testType)
 	return scDir
 }
-func writeJson(data []byte) {
+
+func writeJson(name string, data []byte) {
 	_, basedir, _, ok := runtime.Caller(0)
 	if !ok {
 		panic("no caller info")
@@ -106,9 +145,8 @@ func writeJson(data []byte) {
 	// try to create directory if it doesn't exist
 	_ = os.Mkdir(basedir, os.ModeDir)
 
-	file := filepath.Join(basedir, "tests.json")
-
-	fmt.Printf("writing spec tests json to: %s\n", file)
+	file := filepath.Join(basedir, name+".json")
+	log.Printf("writing spec tests json to: %s\n", file)
 	if err := os.WriteFile(file, data, 0400); err != nil {
 		panic(err.Error())
 	}
